+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Uses [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) & [Semantic Versio
 ### Changed
 - Dependencies have been upgraded.
 - Kebabbble order parser split into seperate dependency project.
+- 'Driver' changed to 'Collector' (internal references unchanged) (#31).
 
 ### Fixed
 - A check made for incoming Slack events was configured incorrectly.